[server] #76 유저하빗 모델 변경

+++ server/models/userhabit.js
@@ -36,7 +36,7 @@ module.exports = (sequelize, DataTypes) => {
         allowNull: false,
       },
       actual_amount: {
-        type: Sequelize.DECIMAL(7, 1),
+        type: Sequelize.INTEGER,
         defaultValue: 0,
         allowNull: false,
       },
